update Le Tableau Noir
After reading my notes, I realized what actually happened was a little different than the narrative I built for the write-up.
I didn't think of red as a signed 32bit int at first, but I did find green which happens to not have the signedness problem in the first place, and that led me to the correct expression for red.
Since this is both more true to reality and makes more sense (why would you be searching for the fail code that displays red vs the win code that displays green and maybe the flag), I rewrote stuff a little bit.